undo resizing of options page
In 2beaee9 the options page size was set to minimum size for proper
display in the Chrome add-on popup context. This also unintentionally
lead to a too-big options page on mobile devices. Because of this the
action was reverted.

Additionally, the "saved" message now has the appropriate background, in
case it's placed over other text.